requirements.txt
requests
gitpython
pyserial
pylint
python-dateutil
